podcast_bot/requirements.txt
httpx>=0.25.0
numpy>=1.24.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
aiogram>=3.0
